refactor(frontend): extract activity card and route painter widgets
